requirements.txt
apache-airflow-providers-postgres==5.14.0
pendulum==3.0.0
psycopg2-binary==2.9.10
requests
pandas
pyarrow